All controller tests are integrated

+++ server/src/main/java/com/example/volunteer_platform/server/service/UserService.java
@@ -51,9 +51,13 @@ public class UserService {
         return userMapper.toUserResponseDTO(user);
     }
 
-    public void deleteUser(String email) {
+    public boolean deleteUser(String email) {
         User user = getUserByEmail(email);
+        if (user == null) {
+            throw new UserNotFoundException("User not found with email: " + email);
+        }
         userRepository.delete(user);
+        return true;
     }
 
     public MessageResponseDTO sendMessage(String messageText, String senderEmail, String recipientEmail) {
